Added Chip Groups and reorganized /features/
