+++ src/App.test.tsx
@@ -6,10 +6,10 @@ import App from './App';
 
 test('renders learn react link', () => {
   const { getByText } = render(
-    <Provider store={store}>
+    <Provider store={ store }>
       <App />
     </Provider>
   );
 
-  expect(getByText(/learn/i)).toBeInTheDocument();
+  expect(getByText(/log/i)).toBeInTheDocument();
 });
