+++ src/app/store.ts
@@ -1,6 +1,6 @@
 import { configureStore, ThunkAction, Action } from "@reduxjs/toolkit"
 import counterReducer from "../features/counter/counterSlice"
-import boardReducer from "../features/sudoku/board/boardSlice"
+import boardReducer from "../features/board/boardSlice"
 
 export const store = configureStore({
   reducer: {
