+++ src/features/chip-group/ChipGroup.module.css
@@ -1,0 +1,6 @@
+.chip-group {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  margin: 0 .5rem
+}
